+++ app/frontend/containers/MapSummaryContainer.js
@@ -9,7 +9,7 @@ const mapStateToProps = state => {
   };
 };
 
-const mapDispatchToProps = (dispatch) => {
+const mapDispatchToProps = (dispatch, ownProps) => {
   return {
   };
 };
